Create login and logout components to make navigation more readable
