+++ src/components/Navigation.js
@@ -1,34 +1,14 @@
 import React from "react";
-import { Navbar, Nav, Button, Form } from "react-bootstrap";
-import { useDispatch, useSelector } from "react-redux";
+import { Navbar, Nav } from "react-bootstrap";
+import { useSelector } from "react-redux";
 import { selectToken } from "../store/user/selectors";
-import { logOut } from "../store/user/actions";
+import LogOutButton from "./LogOutButton";
+import LoginButton from "./LoginButton";
 
 export default function NavigationBar() {
   const token = useSelector(selectToken);
-  const dispatch = useDispatch();
 
-  const loginLogoutButton = token ? (
-    <Nav.Link href="/login">
-      <Form inline>
-        <Button
-          className="mr-sm-2"
-          variant="outline-info"
-          onClick={() => dispatch(logOut())}
-        >
-          Logout
-        </Button>
-      </Form>
-    </Nav.Link>
-  ) : (
-    <Nav.Link href="/login">
-      <Form inline>
-        <Button className="mr-sm-2" variant="outline-info">
-          Login
-        </Button>
-      </Form>
-    </Nav.Link>
-  );
+  const loginLogoutButton = token ? <LogOutButton /> : <LoginButton />;
 
   return (
     <div>
